some code cleaning
