+++ resources/views/dashboard/challenges/refuse_form.blade.php
@@ -10,7 +10,7 @@
 
 
 @section("content")
-	<form action="{{ route("challenges.refuse", ["challengeId" => $challengeId, "teamId" => $teamId]) }}" method="post">
+	<form action="{{ route("validation.refuse", ["challengeId" => $challengeId, "teamId" => $teamId]) }}" method="post">
 	<div class="form-group">
 		<label for="1">Explication du refus</label>
 		<input name="message" id="1" class="form-control" type="text" placeholder="Wallah la photo est nulle.">
