+++ resources/views/layouts/dashboard.blade.php
@@ -50,11 +50,11 @@
 										<ul class="dropdown-menu">
 											@if (Auth::user()->isAdmin())
 												<li><a href="{{ route('challenges.add') }}">Ajouter un défis</a></li>
-												<li><a href={{ route("challenges.validationsList") }}>Liste des validations</a></li>
+												<li><a href={{ route("validation.list") }}>Liste des validations</a></li>
 											@endif
 											<li><a href={{ route('challenges.list') }}>Accéder à la liste des défis</a></li>
-											@if(Auth::user()->ce)
-												<li><a href={{ route("challenges.sent") }}>Défis envoyés pour validation</a></li>
+											@if(Auth::user()->team_id != null)
+												<li><a href={{ route("challenges.sent") }}>Défis relevés </a></li>
 											@endif
 										</ul>
 									</li>
